rename some files and add hygen
