+++ frontend/packages/data-portal/_templates/e2e/new/testFile.cjs.t
@@ -1,15 +1,12 @@
 ---
 to: e2e/<%= name %>.test.ts
 ---
-const hello = ```
-Hello!
-This is your first prompt based hygen template.
+import { test } from '@playwright/test'
+import { <% h.changeCase.pascal(name) %>Page } from 'e2e/pageObjects/<%= name %>/<%= name %>Page'
 
-Learn what it can do here:
-
-https://github.com/jondot/hygen
-```
-
-console.log(hello)
-
-
+test.describe('<%= name %>', () => {
+  test('should work', async ({ page }) => {
+    const <%= name %>Page = new <% h.changeCase.pascal(name) %>Page(page)
+    await <%= name %>Page.goTo('https://playwright.dev/')
+  })
+})
