+++ frontend/packages/data-portal/_templates/e2e/new/types.cjs.t
@@ -1,0 +1,6 @@
+---
+to: e2e/pageObjects/<%= name %>/types.ts
+---
+/**
+  * This file contains test-specific types. Remove if not needed.
+  */
