+++ frontend/packages/data-portal/_templates/e2e/new/util.cjs.t
@@ -1,0 +1,7 @@
+---
+to: e2e/pageObjects/<%= name %>/utils.ts
+---
+/**
+  * This file contains helpers for tests in <%= name %>.test.ts that
+  * are not related to the page itself. For example, data fetchers.
+  */
