+++ frontend/packages/data-portal/e2e/metadataDrawer.test.ts
@@ -7,12 +7,12 @@ import {
   SINGLE_RUN_PATH,
   SINGLE_RUN_URL,
 } from './constants'
-import { MetadataDrawerPage } from './page-objects/metadataDrawer/metadata-drawer-page'
+import { MetadataDrawerPage } from './pageObjects/metadataDrawer/metadataDrawerPage'
 import {
   getAnnotationTestData,
   getSingleDatasetTestMetadata,
   getSingleRunTestMetadata,
-} from './page-objects/metadataDrawer/utils'
+} from './pageObjects/metadataDrawer/utils'
 
 test.describe('Metadata Drawer', () => {
   test.describe(`Single Dataset: ${SINGLE_DATASET_PATH}`, () => {
