+++ frontend/packages/data-portal/e2e/pageObjects/metadataDrawer/utils.ts
